requirements.txt
ai2thor==5.0.0
attrs==21.4.0
aws-requests-auth==0.4.3
black==21.12b0
botocore==1.23.26
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cloudpickle==1.6.0
colour==0.1.5
cycler==0.11.0
datasets==1.18.3
decorator==4.4.2
filelock==3.4.2
Flask==2.0.2
fonttools==4.28.5
future==0.18.2
gym==0.17.3
hydra-core==1.1.1
idna==3.3
imageio==2.13.5
imageio-ffmpeg==0.4.5
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
kiwisolver==1.3.2
llvmlite==0.37.0
MarkupSafe==2.0.1
matplotlib==3.5.1
moviepy==1.0.3
msgpack==1.0.3
mypy-extensions==0.4.3
networkx==2.6.3
numba==0.54.1
numpy #==1.23.1
numpy-quaternion==2021.11.4.15.26.3
opencv-python==4.5.5.62
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.1
proglog==0.1.9
progressbar2==3.55.0
protobuf==3.19.1
pyglet==1.5.0
pyparsing==3.0.6
pyquaternion==0.9.9
python-dateutil==2.8.2
python-utils==2.7.0
python-xlib==0.31
PyYAML==6.0
requests==2.26.0
scipy==1.7.3
setproctitle==1.2.2
six==1.16.0
tensorboardX==2.4.1
tomli==1.2.3
# torch
# torchvision
tqdm==4.62.3
typing-extensions==4.0.1
urllib3==1.26.7
wandb==0.12.13
Werkzeug==2.0.2
shapely==1.8.2
einops
seaborn
pymoo
joblib